new component ConsentItem and removec the button for allowing reminders
permissionsGiven and _reminders are not needed anymore since they only worked as mirrors for the other arrays.
Changes in if statements.
The issue for the checkboxes could not use V-models. Only works with :value
